requirements.txt
requests
piper-tts
soundfile
numpy
